+++ NEWS.md
@@ -1,3 +1,10 @@
+## Features
+
+- Add parameter `timestamp` to make the output of the timestamp optional (#28,
+  #73).
+
+## Internal
+
 - Tests perform strict image comparison (#18).
 - Use `knitr` as vignette builder (#37).